refactor: 모든 노선 역 연결 그래프 생성 시 정상적인 구간 정보만 허용하도록 수정
- 그래프 생성 시 팩토리 패턴 대신 정적 팩토리 메서드 패턴 사용하여 검증 방식 강제함
- 모든 노선 역 연결 그래프 생성 시 노선 역 연결 그래프를 전달받도록 함
- 이를 통해 정상적인 구간 정보만 허용하도록 함
