+++ README.md
@@ -70,6 +70,7 @@
         - [x] 각 변은 어떤 노선에 해당하는 연결 관계인지 알고 있다.
         - [x] 같은 출발지, 도착지를 가진 경로는 1개 이상일 수 있다.
         - [x] 갈래길이 존재할 수 있다.
+        - [x] 단, 각 노선의 역 연결 관계는 단선 그래프와 같은 관계를 가진다. (역 간 경로는 하나만 존재하고, 갈래길이 존재하지 않는다)
 - [ ] 출발역과 도착역을 전달받아 최단 거리 경로를 반환한다.
 - [ ] 한 노선에서 경로 찾기 뿐만 아니라 여러 노선의 환승도 고려한다.
